Add configuration option to enable internal file copying.
Add a new configuration option, --with-cp, that controls whether file copying
by the Mercury compiler is done by invoking the "cp" command or by using the
file copying predicates in the new copy_util module.

When using MSVC or when using MinGW, make the new file copying predicates the
default. (These configurations are the most likely to be the ones where the
Mercury compiler is used from the Windows command prompt, where cp is not
available.)

configure.ac:
    Add the new option, --with-cp.

    Require the bootstrap compiler to support the --install-method configuration
    option that was added last November.

scripts/Mercury.config.in:
     Set the install method configuration option.