Disable the use of asm_fast grades on Windows.
Compilation of the stage 3 library currently results in a segmentation fault on
Windows. Since this can affect users who follow the advice of the configuration
script and attempt to re-install Mercury in a more efficient grades, disable
the use of the asm_fast grades until this is fixed.

configure.ac:
tools/configure_cross:
    As above.