Avoid instances containing clauses in the io module.
The current implementation of intermodule optimization does not allow typeclass
instance definitions containing clauses to occur in .opt files. In particular,
this affects a number of the stream class instances in the io module. This diff
switches over to use 'pred(...) is predname' style of method definition
throughout the io module.

library/io.m:
    Avoid instances that contains clauses; introduce additional private
    predicates where necessary to make this possible.