Replace a long if-then-else chain with switches.
The old code repeated many tests.
The condition of each if-then-else contained several tests, and in many cases,
the initial subsequences were repeated. This meant that if only the later tests
failed, the initial ones would be repeated in the next condition, and the next.
The new code factors all of these out.

compiler/llds_out_data.m:
    When writing out test rvals, instead of testing for one special case
    after another, use a switch on the rval, and then a nested switch
    on the binary operation in binop rvals. There is one remaining
    if-then-else chain, but it is much shorter.

    Use explicit negation flags to avoid almost-duplicated code blocks,
    one of which handles the original and one of which handles the negated
    version of the test.

    Inline and then delete the predicates that previously existed only
    to avoid even more code duplication in these positive/negative pairs.

    Change the code that handles logical_and and logical_or operations
    in ordinary rvals to treat both their arguments as test rvals.