Align functions on word boundaries with clang.
Function trailing with tagged trail entries does not work unless the functions
are aligned on word boundaries. In trailing grades with GCC, we have long
passed an option (-falign-functions) to force this to happen. Do the same for
clang, which supports the same option. This fixes the failure of function
trailing tests on Apple Silicon based machines.
(This does not appear an issue on Intel-based Macs, although passing
-falign-functions option there is harmless.)

scripts/mgnuc.in:
compiler/compile_target_code.m:
     Tell clang to align functions on word boundaries in trailing grades.