Fix a regression in the cast-and-pretest optimization.
For decades now, our automatically generated unification predicates have had
an optimization that casts both inputs to an int and compares the ints.
If an input's value was known, the inst of its cast version was supposed
to be set to the corresponding integer. Somewhere in the last two decades,
our changes to the mode system implementation have broken this optimization,
but this was not detected, because the test case for this problem, which was
added as tests/general/det_complificated_unify2 in 2004, was never enabled.

compiler/modecheck_goal.m:
    Fix this software rot, by allowing the cast values to be in(I) or out(I)
    respectively, instead of insisting on in and out. (In other words, we
    now allow those builtin inst names with arities other than zero.)

compiler/prog_mode.m:
    Add aux predicates needed by new code in modecheck_goal.m.

compiler/hlds_out_goal.m:
    Include a field in the HLDS dump of complicated unifies only if
    it contains useful information.

tests/general/Mmakefile:
    Now that it passes again, enable the test case, which is now called
    det_complificated_unify_2.