Add a string builder that tracks line numbers.
I would like to change mlds_to_c_*.m from writing stuff out directly
to using a string builder. However, the existing string.builder.m
won't work, because when writing out .c/.mh/.mih files with --line-numbers,
we need to generate #line directives before and after the MLDS statements
generated for foreign_procs.

The #line directive before is no problem: the context it writes is
the context of the pragma foreign_proc. The #line directive after, however,
wants to output the line number *in the file we are writing the C code to*.
The usual string builder has no way to peek into the component strings
accumulated so far without destroying its uniqueness. We *could* use
unsafe promises of uniqueness to get around this, but I don't think
that would be a good idea. This is because counting the newline chars
in the component strings so far once for *every* foreign_proc would have
0.5 N*M complexity, where N and M are the number of foreign_procs
and the total length of the generate C file respectively, and for
the kinds of programs we use as stress tests, both can be huge.

It would be better to do in a string builder what the Mercury stream
implementation already does: count the newlines in each string
as it is being written out, and update the current line number
accordingly. This is better because its complexity is linear,
*and* because it has better locality. This diff does that.

compiler/lines_builder.m:
    This new module is a near-copy of string.builder.m. It differs in

    - keeping the (unchanging) filename and the (continually updated)
      line number in the string builder state,

    - adding a filename and an initial line number (usually 1)
      as parameters of the state init function,

    - having a predicate that returns the filename and the line number
      in a given state,

    - to make that predicate possible, threading the state through
      all operations as in/out pairs, not di/uo pairs, and

    - due to that change in modes, NOT having this kind of string builder
      be an instance of the stream typeclasses that the existing
      string builder type is an instance of.

compiler/libs.m:
    Include the new module in this package.

compiler/notes/compiler_design.html:
    Document the new module.