Make --print-error-spec-id apply to extracted msgs.
When e.g. mode analysis prints a diagnostic for a mode error,
it does so by taking an error_spec generated for the *actual* error,
extracing the error_msgs from it, and wrapping a new error_spec around them.
This has meant that if you specify --print-error-spec-id, you get the
id of the predicate that does the wrapping, NOT the id of the predicate
that generated the message describing the actual error.

This diff fixes this.

compiler/error_spec.m:
    Make the predicates that extract error_msgs from error_specs
    also return the error_spec's id.

    Add a version of those predicates that adds a final error_msg
    to the extracted error_msgs that includes this id at the end
    of the message, if --print-error-spec-id is specified.

    To do that, move here the code from write_error_spec.m that
    has always done this job for --print-error-spec-id, and wrap
    an (exported) predicate around it.

compiler/write_error_spec.m:
    Replace the moved code with that predicate.

compiler/mode_errors.m:
compiler/recompilation.check.m:
    Include the id in any extracted error_msgs if asked for.

compiler/error_sort.m:
    Ignore extracted ids.