Ignore singleton arguments in MLDS unifications.
When we generated MLDS code for a deconstruct unification of the form
X = f(Y1, Y2, ..., Yn), we have always generated code that picked up
the values of *all* of the Yi, even the ones that did occur anywhere else.
The code of ml_unused_assign.m then came along and optimized away
the statement that assigned the value of a field to e.g. Y3 if Y3
is not visible outside the unification.

Simply not generating some MLDS code is simpler than generating it
and then later deleting it.

It also directly avoids the creation of some strange code. Given this
if-then-else,

    ( if L = [_H | T], T = [_ | _] then ... else ... )

the compiler used to generate this hlc.gc code for the condition:

    T_6 = ((MR_Word) ((MR_hl_field(1, L_3, (MR_Integer) 1))));
    succeeded = (T_6 != (MR_Word) ((MR_Unsigned) 0U));
    if (succeeded)
    {
    }

It did this because it generated code to pick up the values of *both*
the head and the tail in *both* cons cells, optimized away three of the
resulting four assignment statements, but left the if-then-else around
when it deleted all the statements in the then part.

compiler/ml_unify_gen_util.m:
    Change the utility predicate that decides, given an X = f(Yi) unification,
    which way each argument unification goes between a field of the X term
    and the corresponding variable Yi. If the assignment would notmally be
    to e.g. Y3 but Y3 is not in the unification's nonlocal set, then tell
    our caller that the argument unification is a no-op.

compiler/ml_closure_gen.m:
compiler/ml_code_gen.m:
compiler/ml_unify_gen.m:
compiler/ml_unify_gen_construct.m:
compiler/ml_unify_gen_deconstruct.m:
    Conform to the change in ml_unify_gen_util.m, mostly by passing around
    the nonlocals set to get it to the modified predicate.

compiler/ml_unused_assign.m:
    If we optimize away all statements from both arms of an if-then-else,
    delete the if-then-else altogether.

compiler/mlds_to_c_file.m:
    Unrelated change: fix inconsistent indentation in generated .c files.

compiler/Mercury.options:
    Specify an optimization option that a module updated above
    can benefit from.