Fix an inaccurate error message.
When reporting that a module-qualified imported entity is undefined, we do a
check to see if the module is among those imported and, if not, include an
error message component saying that no module with that name has been imported.
For modules that are ancestors of the module being compiled we should not
emit this error message component since they are always implicitly imported.

compiler/module_qul.qual_errors.m:
    Only emit a "No module named ... has been imported." error message component
    if the module in question is *not* an ancestor of the module being compiled.

tests/invalid_submodules/undef_in_ancestor.{m,err_exp}:
    Add a test for the above.

tests/invalid_submodules/Mmakefile:
    Explain why the above test is not enabled.