Reclassify all the oc_misc options.
Move them to other categories, and add optdb clauses for them.

To make the above possible, create a new category, oc_cmdline,
for the options that modify the command line iself.

Add a new option help_alt. The intention is that it will soon become
the way to test the soon-to-be-implemented new method for generating
help messages from optdb structures. After we are happy with the
results from that, the two help options can switch places, until
we are confident enough to switch off the old system completely.

Separate the "mmc --make" options from the other build system options,
and add optdb clauses for them.