Fix failure of invalid/coerce_typecheck_eqv.m
invalid/coerce_typecheck_eqv.m requires its helper module
coerce_typecheck_eqv_helper_1.m to abstractly export a type that is a
polymorphic equivalence type with a monomorphic definition.
That is not actually supported by the Mercury compiler yet,
but it only reports an error if we try to generate target code for
coerce_typecheck_eqv_helper_1.m (which we don't),
or while generating an intermodule optimisation interface file.

tests/invalid/Mercury.options:
    Disable intermodule optimisation for the test case so as not to
    generate the intermodule optimisation interface file.