Add --mono-path-directory option.
The launcher script we generate for the Mercury compiler in stage 1 does not
work since the CLR does not know where to find the assemblies for the standard
library, mdbcomp etc. This has not been an issue in the past, but it is now a
problem since the doc directory invokes the stage 1 compiler to generate the
invocation section of the user's guide.

If we are using the Mono CLR, then we can tell Mono where to find the library
assemblies by adding the appropriate directories to the value of the MONO_PATH
environment variable we set in the generated launcher script. This diff adds a
new command line option for doing that. Once it has bootstrapped, the
appropriate entries can be added to compiler/COMP_FLAGS.in.

compiler/options.m:
    Add the new option.

compiler/link_target_code.m:
    Include any search directories for Mono in the value of MONO_PATH
    set by the generated launcher script.