Report an error for obviously redundant functional dependencies.
Report an error where the same type variable occurs in both the domain and
range of a functional dependency (e.g. A -> A or A, B -> A). Such dependencies
are redundant, but they are probably not something the programmer intended to
write.

compiler/parse_class.m:
    Report an error where a functional dependency has the same type variable
    in both its domain and range.

tests/invalid_nodepend/Mmakefile:
tests/invalid_nodepend/self_fundep.{m,err_exp}:
    Add a test case for the above error message.