Workaround an issue causing test failures in the C# grade.
A number of test cases are currently failing in the C# grade due to a
ObjectDisposedException being thrown when a binary output file is closed.
This is being caused by the way Mercury files are currently implemented in C#.
In particular, Mercury output streams have both a TextWriter and a
BufferredStream attached to the same underlying IO stream.
The C# implementation of stream_ops.close_stream/4 works by closing the
TextWriter and then closing the BufferedStream. Closing the TextWriter also
closes the IO stream, which triggers the exception when closing the
BufferedStream _unless_ the buffer is empty. The workaround is to flush the
BufferedStream before we close the TextWriter.
(I think something must have changed in, at least, Mono since this was working
previously).

library/io.stream_ops.m:
    Flush the BufferedStream stored in the C# version of a Mercury file before
    closing the TextWriter.

    Add some XXX comments about this and about the way binary streams are
    handled.