Fix library installation with mmake and no symlinks.
When installing various kinds of files (e.g .int*, .opt .mh etc.) for
a library, mmake needs to account for the possibility that a user
of the library may enable --use-subdirs. The legacy library installation
method does this by creating symlinks from within the Mercury directory
back to the locations where the files actually are. On systems that do
not support symlinks (chiefly, non-Cygwin environments on Windows), we copy
the files in question into the appropriate spot under the Mercury
subdirectory.

For .mh files, copying into the Mercury subdirectory is done by the
.install_ints target. This is wrong, as the .mh files are not actually
installed by that target, but are instead installed by the .install_hdrs
target. In non-parallel builds this causes the installation to abort
because there are no .mh files to copy into the Mercury subdirectory.
In parallel builds, the effect is more insidious as some .mh files *can*
be installed by the time .install_ints gets around to copying them and
the copy operation will *not* fail; however, this can result in not
all of the .mh files being copied into the Mercury subdirectory.

The fix is to copy .mh files as part of the .install_hdrs target, not as
part of the .install_ints target.

compiler/generate_mmakefile_fragments.m:
    Do not attempt to symlink the Mercury/mhs directory or copy .mh files
    as part of the .install_ints target. Instead, do it as part of the
    .install_hdrs target.