Let host / target env types be overridden.
Let the host and target env types be overridden in the environment.
If not set in the environment, then they default to the values
determined by the configure script.

Use this capability to let Mercury compilers built in one of the MSYS2
environments be usable from both the MSYS2 shell (via the mmc script) and
also from cmd.exe (via the mercury.bat batch file), _without_ requiring any
modifications to Mercury.config.

scripts/Mercury.config.in:
    Add two new variables that hold the host and target env type.

scripts/mercury.bat.in:
    Set the host and target env types to "windows" when we are
    invoked via this batch file.